+++ apps/dashboard/src/app/(dashboard)/(chain)/components/server/chain-icon.tsx
@@ -1,12 +1,9 @@
 /* eslint-disable @next/next/no-img-element */
 import "server-only";
 
-import {
-  DASHBOARD_THIRDWEB_SECRET_KEY,
-  IPFS_GATEWAY_URL,
-} from "@/constants/env";
+import { thirdwebClient } from "@/constants/client";
+import { DASHBOARD_THIRDWEB_SECRET_KEY } from "@/constants/env";
 import { cn } from "@/lib/utils";
-import { createThirdwebClient } from "thirdweb";
 import { resolveScheme } from "thirdweb/storage";
 
 const fallbackChainIcon =
@@ -17,16 +14,8 @@ export async function ChainIcon(props: {
   className?: string;
 }) {
   if (props.iconUrl) {
-    const client = createThirdwebClient({
-      secretKey: DASHBOARD_THIRDWEB_SECRET_KEY,
-      config: {
-        storage: {
-          gatewayUrl: IPFS_GATEWAY_URL,
-        },
-      },
-    });
     const resolved = resolveScheme({
-      client,
+      client: thirdwebClient,
       uri: props.iconUrl,
     });
     const res = await fetch(resolved, {
