+++ apps/dashboard/src/contexts/custom-sdk-context.tsx
@@ -1,10 +1,10 @@
-import { useQueryClient } from "@tanstack/react-query";
-import { ThirdwebSDKProvider, useSigner } from "@thirdweb-dev/react";
-import type { SDKOptions } from "@thirdweb-dev/sdk";
 import {
   DASHBOARD_THIRDWEB_CLIENT_ID,
   DASHBOARD_THIRDWEB_SECRET_KEY,
-} from "constants/rpc";
+} from "@/constants/env";
+import { useQueryClient } from "@tanstack/react-query";
+import { ThirdwebSDKProvider, useSigner } from "@thirdweb-dev/react";
+import type { SDKOptions } from "@thirdweb-dev/sdk";
 import {
   useSupportedChain,
   useSupportedChains,
