+++ apps/dashboard/src/contract-ui/tabs/manage/page.tsx
@@ -1,12 +1,12 @@
 import { Spinner } from "@/components/ui/Spinner/Spinner";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
+import { thirdwebClient } from "@/constants/client";
 import { useEVMContractInfo } from "@3rdweb-sdk/react";
 import { UserXIcon } from "lucide-react";
 import { useMemo } from "react";
 import { getContract } from "thirdweb";
 import { getInstalledModules, owner } from "thirdweb/extensions/modular";
 import { useActiveAccount, useReadContract } from "thirdweb/react";
-import { thirdwebClient } from "../../../lib/thirdweb-client";
 import { useV5DashboardChain } from "../../../lib/v5-adapter";
 import { InstalledModulesTable } from "./components/InstalledModulesTable";
 import { InstallModuleForm } from "./components/ModuleForm";
