+++ apps/dashboard/src/contract-ui/tabs/nfts/components/airdrop-upload.tsx
@@ -1,3 +1,4 @@
+import { thirdwebClient } from "@/constants/client";
 import {
   AspectRatio,
   Box,
@@ -24,7 +25,6 @@ import {
   VStack,
 } from "@chakra-ui/react";
 import { Logo } from "components/logo";
-import { thirdwebClient } from "lib/thirdweb-client";
 import Papa from "papaparse";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { type DropzoneOptions, useDropzone } from "react-dropzone";
