+++ apps/dashboard/src/contract-ui/tabs/nfts/components/reveal-form.tsx
@@ -1,3 +1,4 @@
+import { thirdwebClient } from "@/constants/client";
 import {
   DrawerBody,
   DrawerFooter,
@@ -17,7 +18,6 @@ import {
 import { TransactionButton } from "components/buttons/TransactionButton";
 import { useTrack } from "hooks/analytics/useTrack";
 import { useTxNotifications } from "hooks/useTxNotifications";
-import { thirdwebClient } from "lib/thirdweb-client";
 import { useV5DashboardChain } from "lib/v5-adapter";
 import { useForm } from "react-hook-form";
 import { getContract } from "thirdweb";
