+++ apps/dashboard/src/lib/defineDashboardChain.ts
@@ -1,6 +1,6 @@
+import { PROD_OR_DEV_URL } from "@/constants/env";
 import { defineChain } from "thirdweb";
 import type { ChainMetadata } from "thirdweb/chains";
-import { PROD_OR_DEV_URL } from "../constants/rpc";
 import { mapV4ChainToV5Chain } from "../contexts/map-chains";
 
 export function defineDashboardChain(
