+++ apps/dashboard/src/lib/sdk.ts
@@ -1,3 +1,9 @@
+import {
+  DASHBOARD_THIRDWEB_CLIENT_ID,
+  DASHBOARD_THIRDWEB_SECRET_KEY,
+  IPFS_GATEWAY_URL,
+  isProd,
+} from "@/constants/env";
 import { type SDKOptions, ThirdwebSDK } from "@thirdweb-dev/sdk";
 import {
   type GatewayUrls,
@@ -6,18 +12,8 @@ import {
   StorageDownloader,
   ThirdwebStorage,
 } from "@thirdweb-dev/storage";
-import {
-  DASHBOARD_THIRDWEB_CLIENT_ID,
-  DASHBOARD_THIRDWEB_SECRET_KEY,
-  isProd,
-} from "constants/rpc";
 import type { Signer } from "ethers";
 import { getAbsoluteUrl } from "./vercel-utils";
-
-// use env var to set IPFS gateway or fallback to ipfscdn.io
-export const IPFS_GATEWAY_URL =
-  (process.env.NEXT_PUBLIC_IPFS_GATEWAY_URL as string) ||
-  "https://{clientId}.ipfscdn.io/ipfs/{cid}/{path}";
 
 export function replaceIpfsUrl(url: string) {
   try {
