--- apps/dashboard/src/lib/thirdweb-client.ts
@@ -1,32 +1,0 @@
-import { DASHBOARD_THIRDWEB_SECRET_KEY } from "@/constants/env";
-import { DASHBOARD_THIRDWEB_CLIENT_ID } from "constants/rpc";
-import { createThirdwebClient } from "thirdweb";
-
-// use env var to set IPFS gateway or fallback to ipfscdn.io
-const IPFS_GATEWAY_URL =
-  (process.env.NEXT_PUBLIC_IPFS_GATEWAY_URL as string) ||
-  "https://{clientId}.ipfscdn.io/ipfs/{cid}/{path}";
-
-export const thirdwebClient = createThirdwebClient(
-  DASHBOARD_THIRDWEB_SECRET_KEY
-    ? {
-        secretKey: DASHBOARD_THIRDWEB_SECRET_KEY,
-        config: {
-          storage: {
-            gatewayUrl: IPFS_GATEWAY_URL,
-          },
-        },
-      }
-    : {
-        clientId: DASHBOARD_THIRDWEB_CLIENT_ID,
-        config: {
-          storage: {
-            gatewayUrl: IPFS_GATEWAY_URL,
-          },
-        },
-      },
-);
-
-/**
- * DO NOT ADD ANYTHING TO THIS FILE IF YOU ARE NOT ABSOLUTELY SURE IT IS OK
- */
