+++ apps/dashboard/src/pages/api/og/contract.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable @next/next/no-img-element */
+import { isProd } from "@/constants/env";
 import { ImageResponse } from "@vercel/og";
-import { isProd } from "constants/rpc";
 import type { NextRequest } from "next/server";
 import { ContractOG } from "og-lib/url-utils";
